+++ internal/config/config_test.go
@@ -6,7 +6,7 @@ import (
 	"testing"
 	"time"
 
-	logger "github.com/soulteary/logger-kit/v2"
+	logger "github.com/soulteary/logger-kit/v3"
 )
 
 func useValidConfig(t *testing.T) {